requirements.txt
fastapi
uvicorn[standard]
python-multipart
pytesseract
Pillow
opencv-python-headless
python-dotenv
reportlab
